requirements.txt
streamlit>=1.18
pymongo>=4.5
pandas>=2.0
geopy>=2.3
folium>=0.14
streamlit-folium>=0.11
python-dotenv>=1.0
